+++ packages/frontend/src/components/head/Preload.tsx
@@ -1,15 +1,23 @@
 import React from 'react'
 
 interface Props {
-  preloadApi?: string
+  preloadApi?: string | string[]
 }
 
 export function Preload(props: Props) {
+  const preloadApi = props.preloadApi ? (
+    Array.isArray(props.preloadApi) ? (
+      props.preloadApi.map((href) => (
+        <link rel="preload" href={href} as="fetch" />
+      ))
+    ) : (
+      <link rel="preload" href={props.preloadApi} as="fetch" />
+    )
+  ) : null
+
   return (
     <>
-      {props.preloadApi && (
-        <link rel="preload" href={props.preloadApi} as="fetch" />
-      )}
+      {preloadApi}
       <link
         rel="preload"
         href="/fonts/Roboto/roboto-v30-latin-regular.woff2"
