+++ packages/frontend/src/scripts/charts/ChartControls.ts
@@ -5,7 +5,7 @@ import { getFilteredSlugs } from '../configureProjectFilters'
 import { getRichSelectValue } from '../configureRichSelect'
 import { makeQuery } from '../query'
 import { setQueryParams } from '../utils/setQueryParams'
-import { ChartDataController } from './ChartDataController'
+import { ChartDataController } from './data-controller/ChartDataController'
 import { ChartSettings, ChartSettingsManager } from './ChartSettings'
 import { ChartType, Milestones, TokenInfo } from './types'
 import { ChartViewController } from './view-controller/ChartViewController'
