+++ packages/frontend/src/scripts/charts/index.ts
@@ -1,6 +1,6 @@
 import { makeQuery } from '../query'
 import { ChartControls } from './ChartControls'
-import { ChartDataController } from './ChartDataController'
+import { ChartDataController } from './data-controller/ChartDataController'
 import { ChartSettingsManager } from './ChartSettings'
 import { ChartRenderer } from './renderer/ChartRenderer'
 import { ChartViewController } from './view-controller/ChartViewController'
